% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_tmhmm.R
\name{install_tmhmm}
\alias{install_tmhmm}
\title{Install tmhmm to a local folder}
\usage{
install_tmhmm(
  download_url = tmhmm::get_tmhmm_url(),
  folder_name = tmhmm::get_default_tmhmm_folder()
)
}
\arguments{
\item{download_url}{the URL to download TMHMM from}

\item{folder_name}{superfolder of TMHMM.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link[tmhmm]{get_default_tmhmm_folder}}
}
\description{
Install tmhmm to a local folder
}
\examples{
if (tmhmm::is_url_valid() && !tmhmm::is_tmhmm_installed()) {
  folder_name <- tempfile()

  install_tmhmm(folder_name = folder_name)

  unlink(folder_name, recursive = TRUE)
}
}
\author{
Richèl J.C. Bilderbeek
}
